requirements.txt
pydantic
pydantic-ai==0.0.24
openai
python-dotenv
okta
sqlalchemy
aiosqlite
pydantic-settings
pytz
fastapi
uvicorn
argon2-cffi
python-jose[cryptography]
python-multipart